Refactoring the implementation
The Tic-Tac-Toe game logic is moved to a separate directory and updated

`jsconfig` was added to the project. This led to the fact that checkJs
in VSCode could be removed because it was no longer needed.

Package type changed to module.
