--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "js/ts.implicitProjectConfig.checkJs": true
-}
