+++ tic-tac-toe/public-api.js
@@ -1,0 +1,2 @@
+export { TicTacToeBoard } from './src/board.js';
+/** @typedef {import('./src/token-utils.js').TicTacToeToken} TicTacToeToken */
